+++ src/test/java/apprenticeship/RulesFor3x3Test.java
